Remove other stuff
