--- examples/atr/vtune_exec.cpp
@@ -1,7 +1,0 @@
-#include <iostream>
-#include <stdlib.h>
-
-int main() {
-    system("python3 simulation.py");
-    return 0;
-}
