+++ examples/pincell/build_xml.py
@@ -68,7 +68,7 @@ settings.particles = 1000
 lower_left = (-pitch/2, -pitch/2, -1)
 upper_right = (pitch/2, pitch/2, 1)
 uniform_dist = openmc.stats.Box(lower_left, upper_right, only_fissionable=True)
-settings.source = openmc.source.Source(space=uniform_dist)
+settings.source = openmc.IndependentSource(space=uniform_dist)
 
 # For source convergence checks, add a mesh that can be used to calculate the
 # Shannon entropy
